vuelvo a double para los montos en vez de BigDecimal

+++ src/main/java/dds/monedero/model/Deposito.java
@@ -5,7 +5,7 @@ import java.time.LocalDate;
 
 public class Deposito extends Movimiento{
 
-  public Deposito(LocalDate fecha, BigDecimal monto) {
+  public Deposito(LocalDate fecha, double monto) {
     super(fecha, monto);
   }
 }
